polymarket-arb-bot/requirements.txt
aiohttp>=3.9.0
asyncio-throttle>=1.0.2
websockets>=12.0
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
chromadb>=0.4.0
rich>=13.7.0
loguru>=0.7.2
pydantic>=2.5.0
schedule>=1.2.0
httpx>=0.25.0
